+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [2.1](../../releases/tag/v2.1)
-- Add Game Over screen sounds and animations based on Sound
+- Improve Game Over screen with sounds and animations
 - Add scaling strategy for components styles
 
 ## [2.0](../../releases/tag/v2.0)